+++ tests/integration_tests/surround.rs
@@ -6,7 +6,7 @@ fn test_surround_single_rect() {
 <rect id="a" xy="0" wh="5" />
 <rect id="s" surround="#a" />
 "##;
-    let expected = r#"<rect id="s" x="0" y="0" width="5" height="5"/>"#;
+    let expected = r#"<rect id="s" x="0" y="0" width="5" height="5" class="d-surround"/>"#;
 
     assert!(transform_str_default(input).unwrap().contains(expected));
 }
@@ -17,7 +17,7 @@ fn test_surround_single_margin() {
 <rect id="a" xy="0" wh="5" />
 <rect id="s" surround="#a" margin="1" />
 "##;
-    let expected = r#"<rect id="s" x="-1" y="-1" width="7" height="7"/>"#;
+    let expected = r#"<rect id="s" x="-1" y="-1" width="7" height="7" class="d-surround"/>"#;
 
     assert!(transform_str_default(input).unwrap().contains(expected));
 }
@@ -30,7 +30,7 @@ fn test_surround_multi_rect() {
 <rect id="c" xy="8" wh="1" />
 <rect id="s" surround="#a #b #c" />
 "##;
-    let expected = r#"<rect id="s" x="0" y="0" width="9" height="12"/>"#;
+    let expected = r#"<rect id="s" x="0" y="0" width="9" height="12" class="d-surround"/>"#;
 
     assert!(transform_str_default(input).unwrap().contains(expected));
 }
@@ -43,7 +43,7 @@ fn test_surround_multi_margin() {
 <rect id="c" xy="8" wh="1" />
 <rect id="s" surround="#a #b #c" margin="1.25 3"/>
 "##;
-    let expected = r#"<rect id="s" x="-1.25" y="-3" width="11.5" height="18"/>"#;
+    let expected = r#"<rect id="s" x="-1.25" y="-3" width="11.5" height="18" class="d-surround"/>"#;
 
     assert!(transform_str_default(input).unwrap().contains(expected));
 }